+++ app/controllers/admin_area/people_controller.rb
@@ -3,10 +3,12 @@ module AdminArea
     # GET /admin/people/1
     def show
       @person        = load_person
-      @spending_info = @person.spending_info
       @account       = @person.account
       @travel_plans  = @account.travel_plans.includes_destinations
       @balances      = @person.balances.includes(:currency)
+      @account_json  = AccountSerializer
+                           .new(@account)
+                           .to_json(include: { people: :spending_info })
 
       card_account_scope = @person.card_accounts.includes(:card, offer: :card)
       @card_accounts = card_account_scope.unpulled
@@ -21,7 +23,7 @@ module AdminArea
     private
 
     def load_person
-      Person.includes(:spending_info).find(params[:id])
+      Person.includes(account: [people: :spending_info]).find(params[:id])
     end
   end
 end